questions/inference_server/model-requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
fastapi==0.75.1
gunicorn==20.1.0
h11==0.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
six==1.14.0
uvicorn==0.17.0
websockets==8.1
wrapt==1.11.2
#transformers==4.26.1 # #not on gae python
#git+https://github.com/huggingface/transformers.git
transformers==4.29.2
tokenizers==0.13.3
#transformers==4.18.0
accelerate==0.19.0

triton
# --find-links https://download.pytorch.org/whl/torch_stable.html
# pip3 install --pre torch torchvision torchaudio --find-links https://download.pytorch.org/whl/nightly/cu121
# torch==1.13.0+cu117 #not on gae python
# torchaudio==0.13.0+cu117
#torch==1.11.0 #not on gae python
#torch==1.10.2
# we dont need torchvision-0.12.0+cu113


google-api-python-client==2.43.0
google-api-core==1.31.5
google-cloud-storage==2.3.0 #not on gae python
#google-cloud-storage==2.0.0

google-cloud-ndb==1.11.1
loguru==0.5.3
pydantic==1.9.0
jinja2==3.1.1 #not on gae python
#jinja2==3.0.3

nltk==3.7.0 #not on gae python
#nltk==3.6.7
stripe==3.4.0
cachetools==4.2.4
python-multipart==0.0.5

# docs hosting
gradio==3.33.1
# OFA link research
bs4==0.0.1
requests==2.27.1
requests-futures==1.0.0

# -e ./OFA/fairseq/ # causes reinstalling a lot
fairseq@git+https://github.com/lee101/fairseq.git
timm
opencv-python==4.6.0.66
ftfy==6.0.3
#tensorboardX==2.4.1
pycocotools==2.0.4
#pycocoevalcap==1.2
pytorch_lightning
einops
datasets
Pillow==9.1.1

numpy==1.22.4
# end OFA

# OCR/pdf processing
pypdf2==2.10.9
pytesseract==0.3.10
setuptools==59.5.0

whisper @ git+https://github.com/openai/whisper.git@15ab54826343c27cfaf44ce31e9c8fb63d0aa775
ffmpeg-python==0.2.0
tensorboard==2.12.0
# download video/audio extract
youtube_dl==2021.12.17
# adding audio/generation
librosa==0.9.2
sentencepiece==0.1.97
